adw.lisp
We will work on this adw.lisp file. this is our agent with the group name on. So take a look at the 1st function . perform-move , what it does is do the foreignAid at the 1st move to take 4 coins, then do the tax move to get 7 coins, when it has 7coins it does coup then coins will go back to 0 , it will do a random action from 7 actions.